// raindrop-backend/src/api/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { ActorNamespace, Ai, Annotation, App, KvCache, Logger, MRNObject, SmartBucket, SmartMemory, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  CHAT_MEMORY: SmartMemory;
  DOCUMENTS: SmartBucket;
  GEMINI_API_KEY: string;
  JWT_SECRET: string;
  logger: Logger;
  mem: KvCache;
  SESSIONS: KvCache;
  tracer: Tracer;
  USER_STATE: ActorNamespace<import('../user-state/index.js').UserState>;
}